+++ app/src/features/card/components/form/CardNumberInput.tsx
@@ -3,7 +3,7 @@ import { CardContext } from "../Card";
 import { CardInput } from "./CardInput";
 import { ErrorMessage } from "./ErrorMessage";
 import { Validator } from "../../validators/CardValidator";
-import { sanitizeErrors } from "../../Utils";
+import { sanitizeErrors } from "../../../../Utils";
 import { CardFieldset, CardLegend } from "../../style/CardStyles";
 
 export function CardNumberInput() {
